requirements.txt
aiocron>=1.8
feedparser>=6.0.10
icalendar>=5.0.11
matrix-nio>=0.23.0
pydeepmerge>=0.3.3
pyyaml>=6.0.1
recurring_ical_events>=2.1.1
safer>=4.10.1
